requirements.txt
Django==1.11.1
oauthlib==1.1.2
PyJWT==1.4.2
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
wheel==0.26.0
MySQL-python
markdown
djangorestframework
django-filter
